unify useAutoScroll
