+++ static/app/components/events/autofix/v3/body.tsx
@@ -1,14 +1,20 @@
-import type {ReactNode} from 'react';
+import type {ReactNode, RefObject, UIEventHandler} from 'react';
 import styled from '@emotion/styled';
 
 import {DrawerBody} from '@sentry/scraps/drawer';
 
 interface SeerDrawerBody {
   children: ReactNode;
+  onScroll?: UIEventHandler;
+  ref?: RefObject<HTMLDivElement | null>;
 }
 
-export function SeerDrawerBody({children}: SeerDrawerBody) {
-  return <StyledDrawerBody>{children}</StyledDrawerBody>;
+export function SeerDrawerBody({children, onScroll, ref}: SeerDrawerBody) {
+  return (
+    <StyledDrawerBody ref={ref} onScroll={onScroll}>
+      {children}
+    </StyledDrawerBody>
+  );
 }
 
 const StyledDrawerBody = styled(DrawerBody)`
